% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{= an integer, indicator of the states to be evaluated, example :c(11,24,36)}

\item{year}{= an integer, the value of the year to be evaluated, example :c(2013,2014,2015)}
}
\value{
This function returns a graph based on the Highway Traffic Fatality data, based in
the years and states selected, the Latitude and Longitud included in the data.
}
\description{
This function takes a year and state numbers and create a graph with the Highway Traffic
Fatality located by longitud and latitude.
The function will provide the warnings and messages included in the fars_read and make_file.
}
\examples{
\dontrun{

fars_map_state(c(20,33,48), c(2013,2014,2015))

}
}
