% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{= a character as the file name of the data in a csv format}
}
\value{
The fars_read function will return a dataframe
}
\description{
This function takes as "input" the name of the file and return a
dataframe using the readr::read_csv function.
}
\note{
In case that the file does not exist will return the message indicating so.
You will need to save the data files in the working directory to avoid error.
}
\examples{
\dontrun{
  fars_read("accident_2013d.csv.bz2")
  }
}
