% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years function}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{= is a vector containing integer values of years to be evaluated like c(2013,2014,2015)}
}
\value{
The fars_read_years returns a data frame with the month and years available in the data
}
\description{
This function takes a vector of years as input, and them select the year and month information contained in
the files (like "accident_YYYYd.csv.bz2") saved in the working directory.
The files will be the ones where the name includes one of the years provided, after appliying
the "make_filename" function.
}
\note{
A warning message will be given in the case that the file or the year and month in the file
are not available.
}
\examples{
\dontrun{

fars_read_years(c(2013,2014,2015))

}
}
