% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{= Is a vector containing the years to be evaluated, like c(2013,2014,2015)}
}
\value{
The function fars_summarize_years returns a data frame with the summarized numbers of month and years
available in the data
}
\description{
This function takes a vector of years, select and summarize the observations by year and month contained
in the files "accident_YYYYd.csv.bz2". The function will provide the warnings and messages included in the
fars_read and fars_read_years functions.
}
\note{
In case that the file does not exist will return the message indicating so.
You will need to save the data files in the working directory.
}
\examples{
\dontrun{

fars_summarize_years (c(2013,2014,2015))

}
}
