% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{= Is the year (integer) of the data that is part of the file name like "accident_2013d.csv.bz2"}
}
\value{
The make_filename function returns the name of the file with the "year" incorporated in the name of the file,
like year=2013 to "accident_2013d.csv.bz2".
}
\description{
This function create a file name based on a given "year" that is part of the file name.
}
\note{
You will need to save the data files in the working directory in order to avoid error.
}
\examples{
\dontrun{
 >example1<-make_filename(2013)
 >example1
 >accident_2013d.csv.bz2
 }
}
